+++ assets/dev/js/frontend/elements-handlers-manager.js
@@ -26,16 +26,12 @@ module.exports = function( $ ) {
 		'wp-widget-media_audio.default': () => import( /* webpackChunkName: 'wp-audio' */ './handlers/wp-audio' ),
 	};
 
-	if ( elementorFrontendConfig.experimentalFeatures[ 'nested-elements' ] && ! elementorFrontendConfig.experimentalFeatures[ 'nested-elements-html' ] ) {
+	if ( elementorFrontendConfig.experimentalFeatures[ 'nested-elements' ] ) {
 		this.elementsHandlers[ 'nested-tabs.default' ] = () => import( /* webpackChunkName: 'nested-tabs' */ 'elementor/modules/nested-tabs/assets/js/frontend/handlers/nested-tabs' );
 	}
 
 	if ( elementorFrontendConfig.experimentalFeatures[ 'nested-elements' ] ) {
 		this.elementsHandlers[ 'nested-accordion.default' ] = () => import( /* webpackChunkName: 'nested-accordion' */ 'elementor/modules/nested-accordion/assets/js/frontend/handlers/nested-accordion' );
-	}
-
-	if ( elementorFrontendConfig.experimentalFeatures[ 'nested-elements-html' ] ) {
-		this.elementsHandlers[ 'nested-tabs-html.default' ] = () => import( /* webpackChunkName: 'nested-tabs-html' */ 'elementor/modules/nested-tabs-html/assets/js/frontend/handlers/nested-tabs-html' );
 	}
 
 	const addGlobalHandlers = () => elementorFrontend.hooks.addAction( 'frontend/element_ready/global', globalHandler );
