+++ assets/dev/js/frontend/modules.js
@@ -6,7 +6,6 @@ import BaseHandler from './handlers/base';
 import SwiperBase from './handlers/base-swiper';
 import CarouselBase from './handlers/base-carousel';
 import NestedTabs from 'elementor/modules/nested-tabs/assets/js/frontend/handlers/nested-tabs';
-import NestedTabsHtml from 'elementor/modules/nested-tabs-html/assets/js/frontend/handlers/nested-tabs-html';
 import NestedAccordion from 'elementor/modules/nested-accordion/assets/js/frontend/handlers/nested-accordion';
 import NestedTitleKeyboardHandler from './handlers/accessibility/nested-title-keyboard-handler';
 
@@ -21,7 +20,6 @@ elementorModules.frontend = {
 		SwiperBase,
 		CarouselBase,
 		NestedTabs,
-		NestedTabsHtml,
 		NestedAccordion,
 		NestedTitleKeyboardHandler,
 	},
