+++ assets/dev/scss/frontend/_modules.scss
@@ -1,5 +1,4 @@
 // Shapes Module
 @import '../../../../modules/shapes/assets/scss/frontend';
 @import '../../../../modules/nested-tabs/assets/scss/frontend';
-@import '../../../../modules/nested-tabs-html/assets/scss/frontend';
 @import '../../../../modules/nested-accordion/assets/scss/frontend';
