--- modules/nested-tabs-html/assets/js/editor/index.js
@@ -1,3 +1,0 @@
-elementorCommon.elements.$window.on( 'elementor/nested-element-type-loaded', async () => {
-	new ( await import( '../editor/module' ) ).default();
-} );
