--- modules/nested-tabs-html/assets/js/editor/module.js
@@ -1,7 +1,0 @@
-import NestedTabsHtml from './nested-tabs-html';
-
-export default class Module {
-	constructor() {
-		elementor.elementsManager.registerElementType( new NestedTabsHtml() );
-	}
-}
