--- modules/nested-tabs-html/assets/js/editor/nested-tabs-html.js
@@ -1,12 +1,0 @@
-import View from './views/view';
-export class NestedTabsHtml extends elementor.modules.elements.types.NestedElementBase {
-	getType() {
-		return 'nested-tabs-html';
-	}
-
-	getView() {
-		return View;
-	}
-}
-
-export default NestedTabsHtml;
