--- modules/nested-tabs-html/module.php
@@ -1,43 +1,0 @@
-<?php
-namespace Elementor\Modules\NestedTabsHtml;
-
-use Elementor\Plugin;
-use Elementor\Modules\NestedElements\Module as NestedElementsModule;
-use Elementor\Core\Experiments\Manager as Experiments_Manager;
-
-if ( ! defined( 'ABSPATH' ) ) {
-	exit; // Exit if accessed directly
-}
-
-class Module extends \Elementor\Core\Base\Module {
-
-	const EXPERIMENT_NAME = 'nested-elements-html';
-
-	public static function get_experimental_data() {
-		return [
-			'name' => self::EXPERIMENT_NAME,
-			'title' => esc_html__( 'Nested Elements HTML', 'elementor' ),
-			'release_status' => Experiments_Manager::RELEASE_STATUS_ALPHA,
-			'default' => Experiments_Manager::STATE_INACTIVE,
-			'hidden' => true,
-		];
-	}
-
-	public static function is_active() {
-		return Plugin::$instance->experiments->is_feature_active( static::EXPERIMENT_NAME );
-	}
-
-	public function get_name() {
-		return 'nested-tabs-html';
-	}
-
-	public function __construct() {
-		parent::__construct();
-
-		add_action( 'elementor/editor/before_enqueue_scripts', function () {
-			wp_enqueue_script( $this->get_name(), $this->get_js_assets_url( $this->get_name() ), [
-				'nested-elements',
-			], ELEMENTOR_VERSION, true );
-		} );
-	}
-}
