+++ modules/nested-tabs/assets/js/editor/nested-tabs.js
@@ -1,6 +1,11 @@
+import View from './views/view';
 export class NestedTabs extends elementor.modules.elements.types.NestedElementBase {
 	getType() {
 		return 'nested-tabs';
+	}
+
+	getView() {
+		return View;
 	}
 }
