+++ modules/nested-tabs/assets/js/editor/views/view.js
