+++ modules/nested-tabs/module.php
@@ -2,7 +2,7 @@
 namespace Elementor\Modules\NestedTabs;
 
 use Elementor\Plugin;
-use  Elementor\Modules\NestedElements\Module as NestedElementsModule;
+use Elementor\Modules\NestedElements\Module as NestedElementsModule;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit; // Exit if accessed directly
