+++ tests/playwright/sanity/modules/nested-tabs/nested-tabs-html.test.ts-snapshots/tabs-direction-left-icon-position-top-align-right-linux.jpeg
