+++ tests/playwright/sanity/modules/nested-tabs/nested-tabs-html.test.ts-snapshots/tabs-direction-right-carousel-visible-linux.jpeg
