+++ tests/playwright/sanity/modules/nested-tabs/templates/tabs-accessibility.json
@@ -1,0 +1,1 @@
+{"content":[{"id":"1d66d846","settings":{"_title":"Container","presetIcon":"eicon-container"},"elements":[{"id":"72570cbd","settings":{"tabs":[{"tab_title":"Tab #1","_id":"d8caae8","tab_icon":{"value":"far fa-address-book","library":"fa-regular"},"tab_icon_active":{"value":"far fa-address-card","library":"fa-regular"}},{"tab_title":"Tab #2","_id":"0daec06","tab_icon":{"value":"far fa-bell","library":"fa-regular"}},{"tab_title":"Tab #3","_id":"dbbb841","tab_icon":{"value":"far fa-calendar-check","library":"fa-regular"},"tab_icon_active":{"value":"far fa-bookmark","library":"fa-regular"}}],"tabs_title_background_color_background":"classic","title_typography_typography":"custom","title_typography_font_family":"Roboto","title_typography_font_weight":"500","title_text_color":"#000000","tabs_title_background_color_hover_color":"#000000","tabs_title_background_color_active_color":"#000000"},"elements":[{"id":"7664bacd","settings":{"_title":"Container","content_width":"full","presetIcon":"eicon-container"},"elements":[{"id":"16559d16","settings":{"title":"Add Your Heading Text Here","title_color":"#000000"},"elements":[],"isInner":false,"widgetType":"heading","elType":"widget"},{"id":"50d5d868","settings":{"text":"Button 1","background_color":"#000000"},"elements":[],"isInner":false,"widgetType":"button","elType":"widget"}],"isInner":false,"elType":"container","isLocked":true},{"id":"697a476f","settings":{"_title":"Container","content_width":"full","presetIcon":"eicon-container"},"elements":[{"id":"7d519f4d","settings":{"text":"Button 2","background_color":"#000000"},"elements":[],"isInner":false,"widgetType":"button","elType":"widget"}],"isInner":false,"elType":"container","isLocked":true},{"id":"50d5d3e1","settings":{"_title":"Container","content_width":"full","presetIcon":"eicon-container"},"elements":[{"id":"351583d","settings":{"text":"Button 3","background_color":"#000000"},"elements":[],"isInner":false,"widgetType":"button","elType":"widget"}],"isInner":false,"elType":"container","isLocked":true}],"isInner":false,"widgetType":"nested-tabs","elType":"widget"}],"isInner":false,"elType":"container"}],"page_settings":[],"version":"0.4","title":"tabs-a11y","type":"container"}
