+++ tests/playwright/sanity/modules/nested-tabs/tests/carousel.ts
@@ -1,24 +1,27 @@
 import { expect, type Page } from '@playwright/test';
+import { selectDropdownContainer, clickTab } from '../helper';
 import EditorPage from '../../../../pages/editor-page';
 import ImageCarousel from '../../../../pages/widgets/image-carousel';
-import { selectDropdownContainer, clickTab } from '../helper';
 
-export async function testCarouselIsVisibleWhenUsingDirectionRightOrLeft( page: Page, editor: EditorPage, imageCarousel: ImageCarousel ) {
+export async function testCarouselIsVisibleWhenUsingDirectionRightOrLeft(
+	page: Page,
+	editor: EditorPage,
+	imageCarousel: ImageCarousel ) {
 	// Act.
-	const contentContainerId = await selectDropdownContainer( editor, '0' ),
-		activeContentContainer = editor.getPreviewFrame().locator( '.e-n-tabs-content > .e-con.e-active' ),
+	const contentContainerId = await selectDropdownContainer( editor, '', 0 ),
+		activeContentContainer = await editor.getPreviewFrame().locator( '.e-n-tabs-content > .e-con.e-active' ),
 		carouselId = await editor.addWidget( 'image-carousel', contentContainerId );
 	// Add images.
 	await imageCarousel.addImageGallery();
 	await imageCarousel.setAutoplay();
 
 	// Set direction right.
-	await clickTab( editor.getPreviewFrame(), '0' );
+	await clickTab( editor.getPreviewFrame(), 0 );
 	await page.locator( '.elementor-control-tabs_direction .eicon-h-align-right' ).click();
 	await editor.togglePreviewMode();
 
 	// Assert
-	expect( await activeContentContainer.screenshot( {
+	expect.soft( await activeContentContainer.screenshot( {
 		type: 'jpeg',
 		quality: 100,
 	} ) ).toMatchSnapshot( 'tabs-direction-right-carousel-visible.jpeg' );
@@ -26,10 +29,6 @@ export async function testCarouselIsVisibleWhenUsingDirectionRightOrLeft( page: 
 	// Restore original view.
 	await editor.togglePreviewMode();
 	await editor.removeElement( carouselId );
-	await clickTab( editor.getPreviewFrame(), '0' );
+	await clickTab( editor.getPreviewFrame(), 0 );
 	await page.locator( '.elementor-control-tabs_direction .eicon-h-align-right' ).click();
 }
-
-module.exports = {
-	testCarouselIsVisibleWhenUsingDirectionRightOrLeft,
-};
