--- tests/playwright/templates/nested-tabs-html-with-icons.json
@@ -1,1 +1,0 @@
-{"content":[{"id":"2052d39b","settings":[],"elements":[{"id":"2d0efe5f","settings":{"tabs":[{"tab_title":"Tab #1","_id":"a7e4e28","tab_icon":{"value":"far fa-arrow-alt-circle-right","library":"fa-regular"},"tab_icon_active":{"value":"far fa-bookmark","library":"fa-regular"}},{"tab_title":"Tab #2","_id":"72f5a56","tab_icon":{"value":"far fa-clipboard","library":"fa-regular"},"tab_icon_active":{"value":"far fa-clock","library":"fa-regular"}},{"tab_title":"Tab #3","_id":"52bb410","tab_icon":{"value":"far fa-clipboard","library":"fa-regular"},"tab_icon_active":{"value":"far fa-address-card","library":"fa-regular"}}],"icon_color_hover":"#FF0000"},"elements":[{"id":"26650f3f","settings":{"_title":"Tab #1"},"elements":[],"isInner":false,"elType":"container","isLocked":true},{"id":"3003375","settings":{"_title":"Tab #2"},"elements":[],"isInner":false,"elType":"container","isLocked":true},{"id":"107f0795","settings":{"_title":"Tab #3"},"elements":[],"isInner":false,"elType":"container","isLocked":true}],"isInner":false,"widgetType":"nested-tabs-html","elType":"widget"}],"isInner":false,"elType":"container"}],"page_settings":[],"version":"0.4","title":"nested-tabs-with-icons","type":"page"}
